feat: hippo4j-config-example rename core to config
